Restructure project to account for distribution as python package
If we would distribute this project as wheel in the way it was before it would pollute
the site-packages as it contained two modules "frontend" and "tazboard". It makes
more sense to have our multiple modules namespaced in "tazboard". So we now have
"tazboard.core" for django boilerplate, "tazboard.frontend" for frontend. New modules
will be placed under tazboard.*

refs #374