Refactor timeframe parameters
It came apparent that using the elastic search notation for dates was impractical to do arithmetic with.
So instead 'now-24h' we actually pass isotime formats which elastic understands as well. Unfortunately that
required some refactoring. Leveraging the new time arithmetics I introduced functions to automatically generate
mocks to demonstrate functioniality without the need to be connected to the actual elastic server. When configured with
TAZBOARD_MOCKS_ENABLED = True, we overwrite the elastic_client.search method to return mocks matching the query instead of
connecting to elastic.

refs #344