Split the complex toplist query into multiple simple ones
Previously, the top X articles (according to clicks) were all requested in one
elastic search aggregation, together with their clicks from the previous time
period (for computing a trend for the clicks), as well as their devices and
referrer. Since this requires huge amounts of in-memory computations on part of
elastic search, we decided to split this one query into multiple simple ones.

Here, we first request the top X articles with their msids and clicks only.
Then, we construct a multi search elastic query, requesting the clicks for the
previous time period, referrers and devices for every msid in the list.