Use updated query for devices and referrer
Important is to aggregate the fingerprints for both, since otherwise
none unique hits were counted. We also attach the msid as a metadata
field to be able to quickly and reliably transform the responses.
Min and max date are not needed anymore since the whole query is pre-filtered
for the corresponding timeframe (in get_interval_exclude_bots)